update: decoraters and imports on flask-jwt-extended

+++ flask-restfull/resources/item.py
@@ -1,5 +1,5 @@
 from  flask_restful  import  Resource ,  reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from modals.item import ItemModel
 
 
@@ -32,8 +32,11 @@ class Item(Resource):
         
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
-        
+       claims = get_jwt_identity()
+       if not claims['is_admin']:
+             return {'message': 'Admin privilege required.'}, 401 
        item = ItemModel.find_by_name(name)
        if item:
           item.delete_from_db()
